backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.14,<5.0
django-cors-headers==4.4.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2,<4.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3,<3.0
drf-yasg==1.21.7
filelock==3.15.4
idna==3.9
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
